add java code, normal and ai.

+++ c/2048-sai.c
@@ -214,13 +214,13 @@ static void init_tables(void) {
         uint8 line[4] = { 0 };
         float score = 0.0f;
 
-        float sum = 0.0f;
+        double sum = 0.0f;
         int empty = 0;
         int merges = 0;
         int prev = 0;
         int counter = 0;
-        float monotonicity_left = 0.0f;
-        float monotonicity_right = 0.0f;
+        double monotonicity_left = 0.0f;
+        double monotonicity_right = 0.0f;
 
         line[0] = (row >> 0) & 0xf;
         line[1] = (row >> 4) & 0xf;
@@ -239,7 +239,7 @@ static void init_tables(void) {
         for (i = 0; i < 4; ++i) {
             int rank = line[i];
 
-            sum += (float)pow(rank, SCORE_SUM_POWER);
+            sum += pow(rank, SCORE_SUM_POWER);
             if (rank == 0) {
                 empty++;
             } else {
@@ -259,10 +259,10 @@ static void init_tables(void) {
         for (i = 1; i < 4; ++i) {
             if (line[i - 1] > line[i]) {
                 monotonicity_left +=
-                    (float)(pow(line[i - 1], SCORE_MONOTONICITY_POWER) - pow(line[i], SCORE_MONOTONICITY_POWER));
+                    pow(line[i - 1], SCORE_MONOTONICITY_POWER) - pow(line[i], SCORE_MONOTONICITY_POWER);
             } else {
                 monotonicity_right +=
-                    (float)(pow(line[i], SCORE_MONOTONICITY_POWER) - pow(line[i - 1], SCORE_MONOTONICITY_POWER));
+                    pow(line[i], SCORE_MONOTONICITY_POWER) - pow(line[i - 1], SCORE_MONOTONICITY_POWER);
             }
         }
